UI design cart and getting the list of product in the cart

+++ app/src/main/java/com/example/tapshopping/data/model/CartResponse.kt
@@ -31,7 +31,7 @@ data class CartResponse (
 
             @SerializedName("cartPrice")
             @Expose
-            var cartPrice: Int? = null,
+            var cartPrice: Double? = null,
 
             @SerializedName("createdAt")
             @Expose
